+++ 규진/프로그래머스/week2/같은_숫자는_싫어.cpp
@@ -1,41 +1,35 @@
-#include <string>
 #include <vector>
-#include <queue>
 #include <iostream>
+#include <stack>
+#include <algorithm>
 
 using namespace std;
 
-vector<int> solution(vector<int> progresses, vector<int> speeds) {
+vector<int> solution(vector<int> arr)
+{
     vector<int> answer;
-    queue<int> q;
-
-    for (int i = 0; i < progresses.size(); i++) {
-        int calNum = (100 - progresses[i]);
-        int num = calNum % speeds[i] == 0 ? calNum / speeds[i] : calNum / speeds[i] + 1;
-        q.push(num);
+    stack<int> st;
+    if (arr.size() <= 0) {
+        return answer;
     }
 
+    st.push(arr[0]);
 
-
-
-    int front = q.front();
-    q.pop();
-    int cnt = 1;
-    while (!q.empty()) {
-        if (front >= q.front()) {
-            q.pop();
-            cnt++;
+    for (int i = 1; i < arr.size(); i++) {
+        int num = st.top();
+        if (num == arr[i]) {
+            continue;
         } else {
-            front = q.front();
-            q.pop();
-            answer.push_back(cnt);
-            cnt = 1;
-        }
-
-        if (q.empty()) {
-            answer.push_back(cnt);
+            st.push(arr[i]);
         }
     }
+
+    while (!st.empty()) {
+        answer.push_back(st.top());
+        st.pop();
+    }
+
+    reverse(answer.begin(), answer.end());
 
     return answer;
 }
